+++ src/EventListener/HookListener.php
@@ -12,12 +12,11 @@
 namespace HeimrichHannot\NewsNavigationBundle\EventListener;
 
 
-use Contao\Controller;
 use Contao\FrontendTemplate;
 use Contao\ModuleNewsReader;
 use Contao\System;
-use HeimrichHannot\Haste\Util\Url;
-use HeimrichHannot\NewsNavigationBundle\Model\NewsModel;
+use HeimrichHannot\NewsNavigationBundle\NewsFilter\FilterFactory;
+use HeimrichHannot\NewsNavigationBundle\NewsFilter\NewsFilter;
 
 class HookListener
 {
@@ -29,10 +28,24 @@ class HookListener
      */
     public function addNewsNavigationLinks ($template, $article, $module)
     {
-        $nextArticle = NewsModel::findNextPublishedByReleaseDate($article['time']);
-        $previousArticle = NewsModel::findPreviousPublishedByReleaseDate($article['time']);
-        $template->nextArticle = $nextArticle->id;
-        $template->previousArticle = $previousArticle->id;
+
+        $filter = System::getContainer()->get('huh.newsnavigation.newsfilter');
+        $filter->filterOnlyPublished();
+        if ($module->newsnavigationRespectArchive)
+        {
+            $filter->filterByArchive($article['pid']);
+        }
+
+        $olderArticle = $filter->createCopy()
+            ->filterOnlyOlderArticles($article['time'])
+            ->getSingleNewsArticle();
+        $newerArticle = $filter->createCopy()
+            ->filterOnlyNewerArticles($article['time'])
+            ->getSingleNewsArticle();
+
+
+        $template->nextArticle = $newerArticle->id;
+        $template->previousArticle = $olderArticle->id;
 
         $template->nextArticleLabel = System::getContainer()->get('translator')->trans('huh.newsnavigation.article.next');
         $template->previousArticleLabel = System::getContainer()->get('translator')->trans('huh.newsnavigation.article.previous');