+++ src/NewsNavigationBundle.php
@@ -12,9 +12,21 @@
 namespace HeimrichHannot\NewsNavigationBundle;
 
 
+use HeimrichHannot\NewsNavigationBundle\DependencyInjection\NewsNavigationExtension;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class NewsNavigationBundle extends Bundle
 {
-
+    /**
+     * Returns the bundle's container extension.
+     *
+     * @throws \LogicException
+     *
+     * @return ExtensionInterface|null The container extension
+     */
+    public function getContainerExtension()
+    {
+        return new NewsNavigationExtension();
+    }
 }